+++ apps/foreldrepengeoversikt/src/components/grupperte-dokumenter/GrupperteDokumenter.tsx
@@ -2,9 +2,10 @@ import { FolderIcon } from '@navikt/aksel-icons';
 
 import { Accordion, BodyShort, HStack } from '@navikt/ds-react';
 
+import { Dokument as DokumentType } from '@navikt/fp-types';
+
 import { guid } from '../../utils/guid';
 import { Dokument } from '../dokument/Dokument';
-import { Dokument as DokumentType } from './../../types/Dokument';
 
 interface Props {
     dokumenter: DokumentType[];
