+++ apps/foreldrepengeoversikt/src/components/har-saker/HarSaker.stories.tsx
@@ -2,7 +2,8 @@ import { Meta, StoryObj } from '@storybook/react/*';
 import { MemoryRouter } from 'react-router-dom';
 import { saker } from 'storybookData/saker/saker';
 
-import { Ytelse } from '../../types/Ytelse';
+import { Ytelse } from '@navikt/fp-types';
+
 import { HarSaker } from './HarSaker';
 
 const meta = {
