+++ apps/foreldrepengeoversikt/src/components/kontonummer-info/KontonummerInfo.tsx
@@ -3,9 +3,7 @@ import { FormattedMessage } from 'react-intl';
 import { Accordion, BodyLong, BodyShort, Button, Detail, Link, VStack } from '@navikt/ds-react';
 
 import { links } from '@navikt/fp-constants';
-import { Bankkonto } from '@navikt/fp-types';
-
-import { Ytelse } from './../../types/Ytelse';
+import { Bankkonto, Ytelse } from '@navikt/fp-types';
 
 interface Props {
     bankkonto: Bankkonto | undefined;
