+++ apps/foreldrepengeoversikt/src/components/minidialog-skjema/minidialogskjemaUtils.ts
@@ -1,9 +1,7 @@
 import { Skjemanummer } from '@navikt/fp-constants';
-import { Attachment } from '@navikt/fp-types';
+import { Attachment, EttersendingDto, Ytelse } from '@navikt/fp-types';
 import { replaceInvisibleCharsWithSpace } from '@navikt/fp-utils';
 
-import { EttersendingDto } from '../../types/EttersendingDTO';
-import { Ytelse } from '../../types/Ytelse';
 import { isAttachmentWithError } from '../../utils/attachmentUtils';
 
 export const mapMinidialogInputTilDTO = (
