+++ apps/foreldrepengeoversikt/src/hooks/useAnnenPartsVedtak.ts
@@ -1,8 +1,9 @@
 import { useQuery } from '@tanstack/react-query';
 
+import { Ytelse } from '@navikt/fp-types';
+
 import { hentAnnenPartsVedtakOptions } from '../api/api';
 import { Sak } from '../types/Sak';
-import { Ytelse } from '../types/Ytelse';
 import { getFamiliehendelseDato } from '../utils/sakerUtils';
 
 export function useAnnenPartsVedtak(sak: Sak | undefined) {
