+++ apps/foreldrepengeoversikt/src/pages/ettersending/EttersendingPage.stories.tsx
@@ -3,9 +3,10 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { HttpResponse, http } from 'msw';
 import { MemoryRouter, Route, Routes } from 'react-router-dom';
 
+import { Ytelse } from '@navikt/fp-types';
+
 import { OversiktRoutes } from '../../routes/routes';
 import { SakOppslag } from '../../types/SakOppslag';
-import { Ytelse } from '../../types/Ytelse';
 import { EttersendingPage } from './EttersendingPage';
 
 const queryClient = new QueryClient({
