+++ apps/foreldrepengeoversikt/src/pages/minidialog-page/MinidialogPage.tsx
@@ -4,6 +4,7 @@ import { useNavigate, useParams } from 'react-router-dom';
 
 import { Heading, VStack } from '@navikt/ds-react';
 
+import { EttersendingDto } from '@navikt/fp-types';
 import { useDocumentTitle } from '@navikt/fp-utils';
 
 import { minidialogOptions, sendEttersending } from '../../api/api';
@@ -15,7 +16,6 @@ import { useSetSelectedRoute } from '../../hooks/useSelectedRoute';
 import { useGetSelectedSak } from '../../hooks/useSelectedSak';
 import { PageRouteLayout } from '../../routes/ForeldrepengeoversiktRoutes';
 import { OversiktRoutes } from '../../routes/routes';
-import { EttersendingDto } from '../../types/EttersendingDTO';
 
 interface Props {
     fnr: string;
