--- apps/foreldrepengeoversikt/src/types/AktørType.ts
@@ -1,5 +1,0 @@
-export enum AktørType {
-    BRUKER = 'BRUKER',
-    NAV = 'NAV',
-    ARBEIDSGIVER = 'ARBEIDSGIVER',
-}
