--- apps/foreldrepengeoversikt/src/types/AnnenPartVedtakDTO.ts
@@ -1,8 +1,0 @@
-import { DekningsgradDTO, SaksperiodeNy } from '@navikt/fp-types';
-
-export interface AnnenPartVedtakDTO {
-    antallBarn?: number;
-    dekningsgrad: DekningsgradDTO;
-    perioder: SaksperiodeNy[];
-    termindato?: string;
-}
