--- apps/foreldrepengeoversikt/src/types/DokumentType.ts
@@ -1,5 +1,0 @@
-export enum DokumentType {
-    'INNGÅENDE_DOKUMENT' = 'INNGÅENDE_DOKUMENT',
-    'UTGÅENDE_DOKUMENT' = 'UTGÅENDE_DOKUMENT',
-    'ARBEIDSGIVER' = 'ARBEIDSGIVER',
-}
