--- apps/foreldrepengeoversikt/src/types/EttersendingDTO.ts
@@ -1,15 +1,0 @@
-import { Attachment } from '@navikt/fp-types';
-
-import { Ytelse } from './Ytelse';
-
-export interface EttersendingDto {
-    brukerTekst?: {
-        dokumentType: string;
-        overskrift: string;
-        tekst: string;
-    };
-    dialogId?: string;
-    saksnummer: string;
-    type: Ytelse;
-    vedlegg: Attachment[];
-}
