+++ apps/foreldrepengeoversikt/src/types/Foreldrepengesak.ts
@@ -1,6 +1,4 @@
-import { ForeldrepengesakDTO } from '@navikt/fp-types';
-
-import { Ytelse } from './Ytelse';
+import { ForeldrepengesakDTO, Ytelse } from '@navikt/fp-types';
 
 export interface Foreldrepengesak extends ForeldrepengesakDTO {
     ytelse: Ytelse.FORELDREPENGER;
