+++ apps/foreldrepengeoversikt/src/types/GruppertSak.ts
@@ -1,9 +1,10 @@
+import { Ytelse } from '@navikt/fp-types';
+
 import { BarnGruppering } from './BarnGruppering';
 import { EngangsstønadSak } from './EngangsstønadSak';
 import { Foreldrepengesak } from './Foreldrepengesak';
 import { Situasjon } from './Situasjon';
 import { SvangerskapspengeSak } from './SvangerskapspengeSak';
-import { Ytelse } from './Ytelse';
 
 export interface GruppertSak {
     antallBarn: number;
