--- apps/foreldrepengeoversikt/src/types/MellomlagredeYtelser.ts
@@ -1,5 +1,0 @@
-export type MellomlagredeYtelser = {
-    engangsstonad: boolean;
-    foreldrepenger: boolean;
-    svangerskapspenger: boolean;
-};
