+++ apps/foreldrepengeoversikt/src/types/SvangerskapspengeSak.ts
@@ -1,6 +1,4 @@
-import { SvangerskapspengeSakDTO } from '@navikt/fp-types';
-
-import { Ytelse } from './Ytelse';
+import { SvangerskapspengeSakDTO, Ytelse } from '@navikt/fp-types';
 
 export interface SvangerskapspengeSak extends SvangerskapspengeSakDTO {
     ytelse: Ytelse.SVANGERSKAPSPENGER;
