--- apps/foreldrepengeoversikt/src/types/TidslinjehendelseType.ts
@@ -1,19 +1,0 @@
-export enum TidslinjehendelseType {
-    FØRSTEGANGSSØKNAD = 'FØRSTEGANGSSØKNAD',
-    FØRSTEGANGSSØKNAD_NY = 'FØRSTEGANGSSØKNAD_NY',
-    ETTERSENDING = 'ETTERSENDING',
-    ENDRINGSSØKNAD = 'ENDRINGSSØKNAD',
-    INNTEKTSMELDING = 'INNTEKTSMELDING',
-    VEDTAK = 'VEDTAK',
-    VENTER_INNTEKTSMELDING = 'VENTER_INNTEKTSMELDING',
-    VENTER_PGA_TIDLIG_SØKNAD = 'VENTER_PGA_TIDLIG_SØKNAD',
-    VENTER_MELDEKORT = 'VENTER_MELDEKORT',
-    VENT_DOKUMENTASJON = 'VENT_DOKUMENTASJON',
-    UTGÅENDE_INNHENT_OPPLYSNINGER = 'UTGÅENDE_INNHENT_OPPLYSNINGER',
-    UTGÅENDE_ETTERLYS_INNTEKTSMELDING = 'UTGÅENDE_ETTERLYS_INNTEKTSMELDING',
-    UTGÅENDE_VARSEL_TILBAKEBETALING = 'UTGÅENDE_VARSEL_TILBAKEBETALING',
-    FAMILIEHENDELSE = 'FAMILIEHENDELSE',
-    BARNET_TRE_ÅR = 'BARNET_TRE_ÅR',
-    FREMTIDIG_VEDTAK = 'FREMTIDIG_VEDTAK',
-    FORELDREPENGER_FEIL_PRAKSIS_UTSETTELSE_INFOBREV = 'FORELDREPENGER_FEIL_PRAKSIS_UTSETTELSE_INFOBREV',
-}
