--- apps/foreldrepengeoversikt/src/types/Ytelse.ts
@@ -1,5 +1,0 @@
-export enum Ytelse {
-    FORELDREPENGER = 'foreldrepenger',
-    SVANGERSKAPSPENGER = 'svangerskapspenger',
-    ENGANGSSTØNAD = 'engangsstønad',
-}
